requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aniso8601==9.0.1
api==0.0.7
async-timeout==4.0.2
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cycler==0.11.0
db-dtypes==1.0.0
decorator==5.1.1
Flask==2.1.1
Flask-RESTful==0.3.9
fonttools==4.31.2
frozenlist==1.3.0
fsspec==2022.2.0
gcsfs==2022.2.0
google-api-core==2.7.1
google-auth==2.6.2
google-auth-oauthlib==0.5.1
google-cloud-bigquery==2.34.2
google-cloud-bigquery-storage==2.13.0
google-cloud-core==2.2.3
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
gspread==5.3.0
gspread-dataframe==3.2.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
JsonForm==0.0.2
jsonschema==4.4.0
JsonSir==0.0.2
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
meteostat==1.6.1
multidict==6.0.2
networkx==2.7.1
nose==1.3.7
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.1
pandas-gbq==0.17.4
Pillow==9.0.1
proto-plus==1.20.3
protobuf==3.19.4
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydata-google-auth==1.4.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
Python-EasyConfig==0.1.7
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
Resource==0.2.1
rsa==4.8
six==1.16.0
tqdm==4.63.1
urllib3==1.26.9
Werkzeug==2.1.1
yarl==1.7.2
zipp==3.7.0
